+++ packages/engine/src/behave-graph/nodes/Profiles/Engine/Values/EntityNodes.ts
@@ -44,8 +44,8 @@ import { getState } from '@etherealengine/hyperflux'
 import { NameComponent } from '@etherealengine/spatial/src/common/NameComponent'
 import { UUIDComponent } from '@etherealengine/spatial/src/common/UUIDComponent'
 import { RigidBodyComponent } from '@etherealengine/spatial/src/physics/components/RigidBodyComponent'
+import { copyTransformToRigidBody } from '@etherealengine/spatial/src/physics/systems/PhysicsPreTransformSystem'
 import { TransformComponent } from '@etherealengine/spatial/src/transform/components/TransformComponent'
-import { copyTransformToRigidBody } from '@etherealengine/spatial/src/transform/systems/TransformSystem'
 import { cloneDeep, isEqual, uniqueId } from 'lodash'
 import { teleportAvatar } from '../../../../../avatar/functions/moveAvatar'
 import { SceneObjectComponent } from '../../../../../scene/components/SceneObjectComponent'
